Add Main Component
